reorganize rings
