+++ README.md
@@ -14,7 +14,7 @@ To factor large integers using Algebraeon
 
 ```
 use std::str::FromStr;
-use algebraeon::{nzq::Natural, rings::number::natural::factorization::factor};
+use algebraeon::{nzq::Natural, rings::rings::natural::factorization::factor};
 
 let n = Natural::from_str("706000565581575429997696139445280900").unwrap();
 let f = factor(n.clone()).unwrap();
