+++ examples/src/lib.rs
@@ -4,7 +4,7 @@ include!(concat!(env!("OUT_DIR"), "/generated_docs.rs"));
 
 #[cfg(test)]
 mod tests {
-    use algebraeon::{rings::number::finite_fields::modulo::Modulo, sets::structure::*};
+    use algebraeon::{rings::rings::finite_fields::modulo::Modulo, sets::structure::*};
 
     #[test]
     fn enumerate_functions_finite_sets() {
