+++ rings/src/lib.rs
@@ -1,6 +1,5 @@
 #![allow(dead_code)]
 pub mod linear;
-pub mod number;
 pub mod polynomial;
 pub mod rings;
 pub mod structure;
