--- rings/src/number/mod.rs
@@ -1,7 +1,0 @@
-pub mod algebraic;
-pub mod anf;
-pub mod finite_fields;
-pub mod integer;
-pub mod natural;
-pub mod num_theory;
-pub mod rational;
