+++ rings/src/rings/finite_fields/conway_polynomials.rs
