+++ rings/src/rings/finite_fields/extension.rs
@@ -56,7 +56,7 @@ where
 }
 
 pub fn f9() -> FieldExtensionByPolynomialQuotientStructure<ModuloCanonicalStructure<3>> {
-    use crate::number::finite_fields::modulo::*;
+    use crate::rings::finite_fields::modulo::*;
     new_finite_field_extension::<ModuloCanonicalStructure<3>>(
         Modulo::<3>::structure().into(),
         Polynomial::from_coeffs(vec![1, 1, 2]),
