+++ rings/src/rings/integer/zimmermann_polys.rs
