+++ rings/src/rings/isolated_algebraic/bisection_gen.rs
