+++ rings/src/rings/isolated_algebraic/mod.rs
