+++ rings/src/rings/isolated_algebraic/padic/isolate.rs
@@ -1,6 +1,6 @@
-use crate::rings::quotient::QuotientStructure;
-
 use super::*;
+use crate::rings::natural::factorization::primes::is_prime;
+use crate::rings::quotient::QuotientStructure;
 use algebraeon_sets::structure::MetaType;
 use std::collections::HashSet;
