+++ rings/src/rings/mod.rs
@@ -1,1 +1,8 @@
+pub mod isolated_algebraic;
+pub mod algebraic_number_fields;
+pub mod finite_fields;
+pub mod integer;
+pub mod natural;
+pub mod num_theory;
+pub mod rational;
 pub mod quotient;
