+++ rings/src/rings/natural/counting.rs
