+++ rings/src/rings/natural/factorization/ecm.rs
@@ -31,7 +31,7 @@ OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
 DAMAGE.
 */
 
-use crate::{number::natural::factorization::primes::is_prime, structure::MetaSemiRing};
+use crate::{rings::natural::factorization::primes::is_prime, structure::MetaSemiRing};
 use algebraeon_nzq::{Natural, Rng, gcd, traits::ModInv};
 use algebraeon_sets::number_theory::primes;
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
