+++ rings/src/rings/natural/functions.rs
